This patch fixes a problem with the NCSA and Combined loggers. Under some circumstances a wrong transferred information value was being written. Fixes bug #885: http://bugs.cherokee-project.com/885
git-svn-id: svn://cherokee-project.com/cherokee/trunk@5155 5dc97367-97f1-0310-9951-d761b3857238

+++ cherokee/connection-protected.h
@@ -171,9 +171,9 @@ struct cherokee_connection {
 
 	/* Traffic
 	 */
-	size_t                        rx;                  /* Bytes received */
+	off_t                         rx;                  /* Bytes received */
 	size_t                        rx_partial;          /* RX partial counter */
-	size_t                        tx;                  /* Bytes sent */
+	off_t                         tx;                  /* Bytes sent */
 	size_t                        tx_partial;          /* TX partial counter */
 	time_t                        traffic_next;        /* Time to update traffic */
